feat: add collaboration keys broadcasting
