+++ src/leader_broadcast.rs
@@ -1,8 +1,18 @@
-use crate::comms_helper::CommsMessageType;
+use crate::comms_helper::{request, serialize_msg, CommsMessageType};
 use crate::errors::BitVMXError;
+use crate::message_queue::MessageQueue;
+use crate::program::participant::CommsAddress;
+use crate::signature_verifier::SignatureVerifier;
+use crate::types::ProgramContext;
+use bitvmx_broker::identification::identifier::Identifier;
 use bitvmx_operator_comms::operator_comms::PubKeyHash;
+use chrono::Utc;
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
+use std::rc::Rc;
+use storage_backend::storage::{KeyValueStore, Storage};
+use tracing::{error, info, warn};
+use uuid::Uuid;
 
 /// Message containing all original messages received by the leader
 #[derive(Debug, Serialize, Deserialize, Clone)]
@@ -107,6 +117,290 @@ impl OriginalMessage {
 
         Ok(())
     }
+}
+
+// ============================================================================
+// Leader Broadcast Helper
+// ============================================================================
+
+/// Helper function to generate storage key for original messages
+fn get_original_messages_key(context_id: &Uuid, msg_type: CommsMessageType) -> String {
+    format!("bitvmx/original_messages/{}/{:?}", context_id, msg_type)
+}
+
+/// Helper for managing leader broadcast functionality
+/// Encapsulates storage, communication, and message queue operations
+pub struct LeaderBroadcastHelper {
+    store: Rc<Storage>,
+}
+
+impl LeaderBroadcastHelper {
+    /// Create a new LeaderBroadcastHelper
+    pub fn new(store: Rc<Storage>) -> Self {
+        Self { store }
+    }
+
+    /// Store an original message received from a non-leader participant
+    /// Messages are stored by context_id (program_id or collaboration_id) and message type
+    pub fn store_original_message(
+        &self,
+        context_id: &Uuid,
+        msg_type: CommsMessageType,
+        original_msg: OriginalMessage,
+    ) -> Result<(), BitVMXError> {
+        let key = get_original_messages_key(context_id, msg_type);
+        let mut messages: Vec<OriginalMessage> = self.store.get(&key)?.unwrap_or_else(|| vec![]);
+
+        // Check if message from this sender already exists
+        if messages
+            .iter()
+            .any(|m| m.sender_pubkey_hash == original_msg.sender_pubkey_hash)
+        {
+            warn!(
+                "Original message from {} already stored for context {} and type {:?}",
+                original_msg.sender_pubkey_hash, context_id, msg_type
+            );
+            return Ok(()); // Don't error, just skip duplicate
+        }
+
+        messages.push(original_msg);
+        self.store.set(&key, messages, None)?;
+        Ok(())
+    }
+
+    /// Get all original messages stored for a given context and message type
+    pub fn get_original_messages(
+        &self,
+        context_id: &Uuid,
+        msg_type: CommsMessageType,
+    ) -> Result<Vec<OriginalMessage>, BitVMXError> {
+        let key = get_original_messages_key(context_id, msg_type);
+        let messages: Vec<OriginalMessage> = self.store.get(&key)?.unwrap_or_else(|| vec![]);
+        Ok(messages)
+    }
+
+    /// Check if all expected messages have been received
+    /// expected_participants should be a list of pubkey_hashes of non-leader participants
+    pub fn has_all_expected_messages(
+        &self,
+        context_id: &Uuid,
+        msg_type: CommsMessageType,
+        expected_participants: &[PubKeyHash],
+    ) -> Result<bool, BitVMXError> {
+        let messages = self.get_original_messages(context_id, msg_type)?;
+        let received_senders: std::collections::HashSet<&PubKeyHash> =
+            messages.iter().map(|m| &m.sender_pubkey_hash).collect();
+
+        for expected in expected_participants {
+            if !received_senders.contains(expected) {
+                return Ok(false);
+            }
+        }
+
+        Ok(true)
+    }
+
+    /// Clear all stored original messages for a given context and message type
+    pub fn clear_original_messages(
+        &self,
+        context_id: &Uuid,
+        msg_type: CommsMessageType,
+    ) -> Result<(), BitVMXError> {
+        let key = get_original_messages_key(context_id, msg_type);
+        self.store.delete(&key)?;
+        Ok(())
+    }
+
+    /// Broadcast stored original messages to all non-leader participants
+    /// This function:
+    /// 1. Retrieves all stored original messages for the context and message type
+    /// 2. Creates a BroadcastedMessage
+    /// 3. Sends it to all non-leader participants
+    /// 4. Clears the stored messages after successful broadcast
+    pub fn broadcast_to_non_leaders(
+        &self,
+        program_context: &ProgramContext,
+        context_id: &Uuid,
+        msg_type: CommsMessageType,
+        non_leader_participants: &[CommsAddress],
+    ) -> Result<(), BitVMXError> {
+        // Get all stored original messages
+        let original_messages = self.get_original_messages(context_id, msg_type)?;
+
+        if original_messages.is_empty() {
+            warn!(
+                "No original messages to broadcast for context {} and type {:?}",
+                context_id, msg_type
+            );
+            return Ok(());
+        }
+
+        // Create BroadcastedMessage
+        let broadcast_timestamp = Utc::now().timestamp_millis();
+        let broadcasted_msg = BroadcastedMessage {
+            original_msg_type: msg_type,
+            original_messages: original_messages.clone(),
+            broadcast_timestamp,
+        };
+
+        // Validate the broadcasted message
+        broadcasted_msg.validate()?;
+
+        // Send to all non-leader participants
+        for participant in non_leader_participants {
+            info!("Sending leader aggregated key to peer: {:?}", participant);
+
+            request(
+                &program_context.comms,
+                &program_context.key_chain,
+                context_id,
+                participant.clone(),
+                CommsMessageType::Broadcasted,
+                &broadcasted_msg,
+            )?;
+        }
+
+        // Clear stored messages after successful broadcast
+        self.clear_original_messages(context_id, msg_type)?;
+
+        info!(
+            "Successfully broadcasted {} messages to {} non-leaders for context {} and type {:?}",
+            original_messages.len(),
+            non_leader_participants.len(),
+            context_id,
+            msg_type
+        );
+
+        Ok(())
+    }
+
+    /// Process a BroadcastedMessage by recursively processing each original message
+    /// This function:
+    /// 1. Deserializes the BroadcastedMessage
+    /// 2. Validates the BroadcastedMessage structure
+    /// 3. Verifies the leader's signature (already done in process_msg)
+    /// 4. For each OriginalMessage:
+    ///    - Verifies the original message signature
+    ///    - Reconstructs the serialized message
+    ///    - Queues the message for processing
+    pub fn process_broadcasted_message(
+        &self,
+        program_context: &ProgramContext,
+        leader_identifier: Identifier,
+        program_id: Uuid,
+        data: Value,
+        message_queue: &MessageQueue,
+    ) -> Result<(), BitVMXError> {
+        // Deserialize BroadcastedMessage from data
+        let broadcasted_msg: BroadcastedMessage =
+            serde_json::from_value(data.clone()).map_err(|e| {
+                error!("Failed to deserialize BroadcastedMessage: {:?}", e);
+                BitVMXError::InvalidMessageFormat
+            })?;
+
+        // Validate the BroadcastedMessage structure
+        broadcasted_msg.validate()?;
+
+        info!(
+            "Processing BroadcastedMessage from leader {} for context {} with {} original messages",
+            leader_identifier.pubkey_hash,
+            program_id,
+            broadcasted_msg.original_messages.len()
+        );
+
+        // Process each original message recursively
+        for original_msg in &broadcasted_msg.original_messages {
+            // Verify the original message signature
+            let original_verified = Self::verify_original_message_signature(
+                program_context,
+                &original_msg.sender_pubkey_hash,
+                &program_id,
+                &original_msg.version,
+                &original_msg.msg_type,
+                &original_msg.data,
+                original_msg.original_timestamp,
+                &original_msg.original_signature,
+            )?;
+
+            if !original_verified {
+                warn!(
+                    "Original message from {} failed signature verification, skipping",
+                    original_msg.sender_pubkey_hash
+                );
+                continue;
+            }
+
+            // Reconstruct the full serialized message from OriginalMessage
+            let full_message = serialize_msg(
+                &original_msg.version,
+                original_msg.msg_type,
+                &program_id,
+                &original_msg.data,
+                original_msg.original_timestamp,
+                original_msg.original_signature.clone(),
+            )?;
+
+            info!(
+                "Pending message to back: {:?} from {}",
+                original_msg.msg_type, original_msg.sender_pubkey_hash
+            );
+            message_queue.push_back(
+                Identifier::new(original_msg.sender_pubkey_hash.clone(), 0).to_string(),
+                full_message,
+            )?;
+        }
+
+        info!(
+            "Successfully queued BroadcastedMessage from leader {} with {} original messages",
+            leader_identifier.pubkey_hash,
+            broadcasted_msg.original_messages.len()
+        );
+
+        Ok(())
+    }
+
+    /// Verify the signature of an original message
+    /// This is similar to verify_message_signature but works with OriginalMessage data
+    fn verify_original_message_signature(
+        program_context: &ProgramContext,
+        sender_pubkey_hash: &PubKeyHash,
+        program_id: &Uuid,
+        version: &String,
+        msg_type: &CommsMessageType,
+        data: &Value,
+        timestamp: i64,
+        signature: &Vec<u8>,
+    ) -> Result<bool, BitVMXError> {
+        match SignatureVerifier::verify_and_get_key(
+            &program_context.comms,
+            &program_context.globals,
+            &program_context.key_chain,
+            sender_pubkey_hash,
+            program_id,
+            msg_type,
+            data,
+            timestamp,
+            signature,
+            version,
+        ) {
+            Ok(_) => Ok(true),
+            Err(BitVMXError::MissingVerificationKey { .. }) => Ok(false),
+            Err(err) => Err(err),
+        }
+    }
+}
+
+/// Get list of non-leader participants from a list of all participants
+/// Returns CommsAddress of all participants except the leader
+pub fn get_non_leader_participants(
+    all_participants: &[CommsAddress],
+    leader_pubkey_hash: &PubKeyHash,
+) -> Vec<CommsAddress> {
+    all_participants
+        .iter()
+        .filter(|p| &p.pubkey_hash != leader_pubkey_hash)
+        .cloned()
+        .collect()
 }
 
 #[cfg(test)]
